+++ src/main/java/com/mongodb/kafka/connect/sink/writemodel/strategy/DeleteOneBusinessKeyStrategy.java
@@ -12,28 +12,31 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
- *
- * Original Work: Apache License, Version 2.0, Copyright 2017 Hans-Peter Grahsl.
  */
 
 package com.mongodb.kafka.connect.sink.writemodel.strategy;
 
 import static com.mongodb.kafka.connect.sink.MongoSinkTopicConfig.ID_FIELD;
+import static com.mongodb.kafka.connect.sink.writemodel.strategy.WriteModelHelper.flattenKeys;
 
 import org.apache.kafka.connect.errors.DataException;
 
 import org.bson.BsonDocument;
 import org.bson.BsonValue;
 
-import com.mongodb.client.model.ReplaceOneModel;
-import com.mongodb.client.model.ReplaceOptions;
+import com.mongodb.client.model.DeleteOneModel;
 import com.mongodb.client.model.WriteModel;
 
+import com.mongodb.kafka.connect.sink.Configurable;
+import com.mongodb.kafka.connect.sink.MongoSinkTopicConfig;
 import com.mongodb.kafka.connect.sink.converter.SinkDocument;
+import com.mongodb.kafka.connect.sink.processor.id.strategy.IdStrategy;
+import com.mongodb.kafka.connect.sink.processor.id.strategy.PartialKeyStrategy;
+import com.mongodb.kafka.connect.sink.processor.id.strategy.PartialValueStrategy;
 
-public class ReplaceOneDefaultStrategy implements WriteModelStrategy {
+public class DeleteOneBusinessKeyStrategy implements WriteModelStrategy, Configurable {
 
-  private static final ReplaceOptions REPLACE_OPTIONS = new ReplaceOptions().upsert(true);
+  private boolean isPartialId = false;
 
   @Override
   public WriteModel<BsonDocument> createWriteModel(final SinkDocument document) {
@@ -46,11 +49,24 @@ public class ReplaceOneDefaultStrategy implements WriteModelStrategy {
                         "Could not build the WriteModel,the value document was missing unexpectedly"));
 
     BsonValue idValue = vd.get(ID_FIELD);
-    if (idValue == null) {
+    if (idValue == null || !idValue.isDocument()) {
       throw new DataException(
-          "Could not build the WriteModel,the `_id` field was missing unexpectedly");
+          "Could not build the WriteModel, the value document does not contain an _id field of"
+              + " type BsonDocument which holds the business key fields.");
     }
 
-    return new ReplaceOneModel<>(new BsonDocument(ID_FIELD, idValue), vd, REPLACE_OPTIONS);
+    BsonDocument businessKey = idValue.asDocument();
+    vd.remove(ID_FIELD);
+    if (isPartialId) {
+      businessKey = flattenKeys(businessKey);
+    }
+    return new DeleteOneModel<>(businessKey);
+  }
+
+  @Override
+  public void configure(final MongoSinkTopicConfig configuration) {
+    IdStrategy idStrategy = configuration.getIdStrategy();
+    isPartialId =
+        idStrategy instanceof PartialKeyStrategy || idStrategy instanceof PartialValueStrategy;
   }
 }