+++ src/main/java/com/mongodb/kafka/connect/sink/writemodel/strategy/ReplaceOneBusinessKeyStrategy.java
@@ -23,8 +23,8 @@ import static com.mongodb.kafka.connect.sink.writemodel.strategy.WriteModelHelpe
 
 import org.apache.kafka.connect.errors.DataException;
 
-import org.bson.BSONException;
 import org.bson.BsonDocument;
+import org.bson.BsonValue;
 
 import com.mongodb.client.model.ReplaceOneModel;
 import com.mongodb.client.model.ReplaceOptions;
@@ -52,19 +52,20 @@ public class ReplaceOneBusinessKeyStrategy implements WriteModelStrategy, Config
                     new DataException(
                         "Could not build the WriteModel,the value document was missing unexpectedly"));
 
-    try {
-      BsonDocument businessKey = vd.getDocument(ID_FIELD);
-      vd.remove(ID_FIELD);
-      if (isPartialId) {
-        businessKey = flattenKeys(businessKey);
-      }
-      return new ReplaceOneModel<>(businessKey, vd, REPLACE_OPTIONS);
-    } catch (BSONException e) {
+    BsonValue idValue = vd.get(ID_FIELD);
+    if (idValue == null || !idValue.isDocument()) {
       throw new DataException(
           "Could not build the WriteModel,the value document does not contain an _id field of"
               + " type BsonDocument which holds the business key fields.\n\n If you are including an"
               + " existing `_id` value in the business key then ensure `document.id.strategy.overwrite.existing=true`.");
     }
+
+    BsonDocument businessKey = idValue.asDocument();
+    vd.remove(ID_FIELD);
+    if (isPartialId) {
+      businessKey = flattenKeys(businessKey);
+    }
+    return new ReplaceOneModel<>(businessKey, vd, REPLACE_OPTIONS);
   }
 
   @Override